requirements.txt
boto3
azure-storage-blob
requests
